Tagger can now add songs to DB and GUI can read from DB.  Seekbar now updates and seeks correctly.

+++ src/gui/Mp3PositionListener.java
@@ -8,8 +8,10 @@ public interface Mp3PositionListener {
 	
 	/*
 	 * called when the position of the song has changed
+	 * 
+	 * @param pos position of track 0.0 - 1.0
 	 */
-	void updateSeektime(long pos);
+	void updateSeektime(double pos);
 	
 	/*
 	 * called when the song has completed playing (not called
